+++ src/db_options.rs
@@ -161,6 +161,13 @@ impl Cache {
         Cache(Arc::new(CacheWrapper { inner }))
     }
 
+    /// Creates an LRU cache with custom options.
+    pub fn new_lru_cache_opts(opts: &LruCacheOptions) -> Cache {
+        let inner =
+            NonNull::new(unsafe { ffi::rocksdb_cache_create_lru_opts(opts.inner) }).unwrap();
+        Cache(Arc::new(CacheWrapper { inner }))
+    }
+
     /// Creates a HyperClockCache with capacity in bytes.
     ///
     /// `estimated_entry_charge` is an important tuning parameter. The optimal
@@ -291,7 +298,11 @@ pub struct Options {
 /// ```
 /// use rocksdb::{DB, Options, WriteBatch, WriteOptions};
 ///
-/// let path = "_path_for_rocksdb_storageY1";
+/// let tempdir = tempfile::Builder::new()
+///     .prefix("_path_for_rocksdb_storageY1")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_rocksdb_storageY1");
+/// let path = tempdir.path();
 /// {
 ///     let db = DB::open_default(path).unwrap();
 ///     let mut batch = WriteBatch::default();
@@ -311,6 +322,10 @@ pub struct WriteOptions {
     pub(crate) inner: *mut ffi::rocksdb_writeoptions_t,
 }
 
+pub struct LruCacheOptions {
+    pub(crate) inner: *mut ffi::rocksdb_lru_cache_options_t,
+}
+
 /// Optionally wait for the memtable flush to be performed.
 ///
 /// # Examples
@@ -320,7 +335,11 @@ pub struct WriteOptions {
 /// ```
 /// use rocksdb::{DB, Options, FlushOptions};
 ///
-/// let path = "_path_for_rocksdb_storageY2";
+/// let tempdir = tempfile::Builder::new()
+///     .prefix("_path_for_rocksdb_storageY2")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_rocksdb_storageY2");
+/// let path = tempdir.path();
 /// {
 ///     let db = DB::open_default(path).unwrap();
 ///
@@ -369,18 +388,26 @@ pub struct CuckooTableOptions {
 ///
 /// let writer_opts = Options::default();
 /// let mut writer = SstFileWriter::create(&writer_opts);
-/// writer.open("_path_for_sst_file").unwrap();
+/// let tempdir = tempfile::Builder::new()
+///     .tempdir()
+///     .expect("Failed to create temporary folder for the _path_for_sst_file");
+/// let path1 = tempdir.path().join("_path_for_sst_file");
+/// writer.open(path1.clone()).unwrap();
 /// writer.put(b"k1", b"v1").unwrap();
 /// writer.finish().unwrap();
 ///
-/// let path = "_path_for_rocksdb_storageY3";
+/// let tempdir2 = tempfile::Builder::new()
+///     .prefix("_path_for_rocksdb_storageY3")
+///     .tempdir()
+///     .expect("Failed to create temporary path for the _path_for_rocksdb_storageY3");
+/// let path2 = tempdir2.path();
 /// {
-///   let db = DB::open_default(&path).unwrap();
+///   let db = DB::open_default(&path2).unwrap();
 ///   let mut ingest_opts = IngestExternalFileOptions::default();
 ///   ingest_opts.set_move_files(true);
-///   db.ingest_external_file_opts(&ingest_opts, vec!["_path_for_sst_file"]).unwrap();
+///   db.ingest_external_file_opts(&ingest_opts, vec![path1]).unwrap();
 /// }
-/// let _ = DB::destroy(&Options::default(), path);
+/// let _ = DB::destroy(&Options::default(), path2);
 /// ```
 pub struct IngestExternalFileOptions {
     pub(crate) inner: *mut ffi::rocksdb_ingestexternalfileoptions_t,
@@ -391,6 +418,7 @@ pub struct IngestExternalFileOptions {
 // rocksdb internally does not rely on thread-local information for its user-exposed types.
 unsafe impl Send for Options {}
 unsafe impl Send for WriteOptions {}
+unsafe impl Send for LruCacheOptions {}
 unsafe impl Send for FlushOptions {}
 unsafe impl Send for BlockBasedOptions {}
 unsafe impl Send for CuckooTableOptions {}
@@ -404,6 +432,7 @@ unsafe impl Send for WriteBufferManagerWrapper {}
 // use within the rocksdb library is generally behind a const reference
 unsafe impl Sync for Options {}
 unsafe impl Sync for WriteOptions {}
+unsafe impl Sync for LruCacheOptions {}
 unsafe impl Sync for FlushOptions {}
 unsafe impl Sync for BlockBasedOptions {}
 unsafe impl Sync for CuckooTableOptions {}
@@ -461,6 +490,14 @@ impl Drop for WriteOptions {
     fn drop(&mut self) {
         unsafe {
             ffi::rocksdb_writeoptions_destroy(self.inner);
+        }
+    }
+}
+
+impl Drop for LruCacheOptions {
+    fn drop(&mut self) {
+        unsafe {
+            ffi::rocksdb_lru_cache_options_destroy(self.inner);
         }
     }
 }
@@ -2239,6 +2276,29 @@ impl Options {
         }
     }
 
+    /// Sets the compaction priority. When multiple files are picked for compaction from a level,
+    /// this option determines which files to pick first.
+    ///
+    /// Default: `CompactionPri::ByCompensatedSize`
+    ///
+    /// Dynamically changeable through SetOptions() API
+    ///
+    /// See [rocksdb post](https://github.com/facebook/rocksdb/blob/f20d12adc85ece3e75fb238872959c702c0e5535/docs/_posts/2016-01-29-compaction_pri.markdown) for more details.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use rocksdb::{Options, CompactionPri};
+    ///
+    /// let mut opts = Options::default();
+    /// opts.set_compaction_pri(CompactionPri::MinOverlappingRatio);
+    /// ```
+    pub fn set_compaction_pri(&mut self, pri: CompactionPri) {
+        unsafe {
+            ffi::rocksdb_options_set_compaction_pri(self.inner, pri as i32);
+        }
+    }
+
     /// Sets the soft limit on number of level-0 files. We start slowing down writes at this
     /// point. A value < `0` means that no writing slow down will be triggered by
     /// number of files in level-0.
@@ -3609,6 +3669,39 @@ impl Default for WriteOptions {
     }
 }
 
+impl LruCacheOptions {
+    /// Capacity of the cache, in the same units as the `charge` of each entry.
+    /// This is typically measured in bytes, but can be a different unit if using
+    /// kDontChargeCacheMetadata.
+    pub fn set_capacity(&mut self, cap: usize) {
+        unsafe {
+            ffi::rocksdb_lru_cache_options_set_capacity(self.inner, cap);
+        }
+    }
+
+    /// Cache is sharded into 2^num_shard_bits shards, by hash of key.
+    /// If < 0, a good default is chosen based on the capacity and the
+    /// implementation. (Mutex-based implementations are much more reliant
+    /// on many shards for parallel scalability.)
+    pub fn set_num_shard_bits(&mut self, val: c_int) {
+        unsafe {
+            ffi::rocksdb_lru_cache_options_set_num_shard_bits(self.inner, val);
+        }
+    }
+}
+
+impl Default for LruCacheOptions {
+    fn default() -> Self {
+        let inner = unsafe { ffi::rocksdb_lru_cache_options_create() };
+        assert!(
+            !inner.is_null(),
+            "Could not create RocksDB LRU cache options"
+        );
+
+        Self { inner }
+    }
+}
+
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
 #[repr(i32)]
@@ -3621,6 +3714,27 @@ pub enum ReadTier {
     Persisted,
     /// Reads data in memtable. Used for memtable only iterators.
     Memtable,
+}
+
+#[repr(i32)]
+pub enum CompactionPri {
+    /// Slightly prioritize larger files by size compensated by #deletes
+    ByCompensatedSize = 0,
+    /// First compact files whose data's latest update time is oldest.
+    /// Try this if you only update some hot keys in small ranges.
+    OldestLargestSeqFirst = 1,
+    /// First compact files whose range hasn't been compacted to the next level
+    /// for the longest. If your updates are random across the key space,
+    /// write amplification is slightly better with this option.
+    OldestSmallestSeqFirst = 2,
+    /// First compact files whose ratio between overlapping size in next level
+    /// and its size is the smallest. It in many cases can optimize write amplification.
+    MinOverlappingRatio = 3,
+    /// Keeps a cursor(s) of the successor of the file (key range) was/were
+    /// compacted before, and always picks the next files (key range) in that
+    /// level. The file picking process will cycle through all the files in a
+    /// round-robin manner.
+    RoundRobin = 4,
 }
 
 impl ReadOptions {
@@ -4466,7 +4580,7 @@ impl Drop for DBPath {
 #[cfg(test)]
 mod tests {
     use crate::db_options::WriteBufferManager;
-    use crate::{Cache, MemtableFactory, Options};
+    use crate::{Cache, CompactionPri, MemtableFactory, Options};
 
     #[test]
     fn test_enable_statistics() {
@@ -4517,5 +4631,30 @@ mod tests {
 
         // WriteBufferManager outlives options
         assert!(write_buffer_manager.enabled());
+    }
+
+    #[test]
+    fn compaction_pri() {
+        let mut opts = Options::default();
+        opts.set_compaction_pri(CompactionPri::RoundRobin);
+        opts.create_if_missing(true);
+        let tmp = tempfile::tempdir().unwrap();
+        let _db = crate::DB::open(&opts, tmp.path()).unwrap();
+
+        let options = std::fs::read_dir(tmp.path())
+            .unwrap()
+            .find_map(|x| {
+                let x = x.ok()?;
+                x.file_name()
+                    .into_string()
+                    .unwrap()
+                    .contains("OPTIONS")
+                    .then_some(x.path())
+            })
+            .map(std::fs::read_to_string)
+            .unwrap()
+            .unwrap();
+
+        assert!(options.contains("compaction_pri=kRoundRobin"));
     }
 }