+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## [Unreleased]
 * Document that `default` column family doesn't inherit open options of db (
   0xdeafbeef)
+* Expose LRU cache options (athre0z)
+* Add `with_capacity_bytes` to `WriteBatch` (0xdeafbeef)
+* Add `set_compaction_pri` to `Options` (0xdeafbeef)
+
+## [Breaking Changes]
+* Update jemalloc-sys to 0.6.0 (0xdeafbeef)
+* Bump MSRV to 1.71.1 (0xdeafbeef)
 
 ## 0.22.0 (2024-02-13)